crypto: move interactive verification tests to a new file
